# orders-microservice/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: YumYumNow Orders API
  version: 1.0.0
  description: API for handling orders in the YumYumNow application.
servers:
  - url: 'https://localhost:8082'
...
paths:
...
  /vendor/{vendorId}/dish:
    post:
      tags:
        - Vendor
      summary: Add a new dish to the vendor menu
      operationId: addDishToMenu
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dish'
            examples:
              example1:
                summary: Example dish
                value:
                  name: "Margherita Pizza"
                  imageLink: "http://example.com/images/margherita.jpg"
                  price: 12.50
                  allergens: ["Gluten", "Dairy"]
                  ingredients: ["Flour", "Tomatoes", "Mozzarella", "Basil", "Olive Oil"]
                  description: "Best pizza in Delft!"
      responses:
        '200':
          description: New dish has been added to the vendor menu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
              examples:
                exampleResponse:
                  summary: Example response of a newly added dish
                  value:
                    ID: "ccf1f2d7-8f7f-4044-9a30-16790a432977"
                    name: "Margherita Pizza"
                    imageLink: "http://example.com/images/margherita.jpg"
                    price: 12.50
                    allergens: ["Gluten", "Dairy"]
                    ingredients: ["Flour", "Tomatoes", "Mozzarella", "Basil", "Olive Oil"]
                    description: "Best pizza in Delft!"
        '400':
          description: Bad Request - Incorrect dish details format or missing required fields.
        '401':
          description: Unauthorized - User is not a vendor
        '404':
          description: Vendor not found
        '500':
          description: Internal Server Error - An unexpected error occurred on the server.
  /vendor/{vendorId}/orders:
    get:
      tags:
        - Vendor
      summary: Get orders for a vendor
      description: A vendor can ask for all of their orders. They only receive orders that have been paid for, so the ones with status "accepted".
      operationId: getVendorOrders
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        
      responses:
        '200':
          description: List of orders for the vendor
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '400':
          description:  Bad Request - Invalid request parameters. 
        '401':
          description: Unauthorized - User is not a vendor          
        '404':
          description: Vendor not found
        '500':
          description: Internal Server Error - An unexpected error occurred on the server.
...
  /vendor/{vendorId}/dish/{dishId}:
    get:
      tags:
        - Vendor
      summary: Get details of a specific dish
      description: Get detailed information about a specific dish offered by the vendor.
      operationId: getDish
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: dishId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detailed information about the specified dish.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
              examples:
                example1:
                  summary: Example dish details
                  value:
                    ID: "ccf1f2d7-8f7f-4044-9a30-16790a432977"
                    name: "Margherita Pizza"
                    imageLink: "http://example.com/images/margherita.jpg"
                    price: 12.50
                    allergens: ["Gluten", "Dairy"]
                    ingredients: ["Flour", "Tomatoes", "Mozzarella", "Basil", "Olive Oil"]
                    description: "Best pizza in Delft!"
        '400':
          description:  Bad Request - Invalid request parameters.               
        '401':
          description: Unauthorized - User is a vendor/dish does not belong to vendor.
        '404':
          description: Dish or vendor not found.
        '500':
          description: Internal Server Error - An unexpected error occurred on the server.

    delete:
      tags:
        - Vendor
      summary: Delete a dish from the menu
      operationId: removeDishFromMenu
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: dishId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Dish removed successfully from the menu.
        '400':
          description: Bad Request - Dish not deleted.
        '401':
          description: Unauthorized - User is not a vendor/dish does not belong to vendor.
        '404':
          description: Not Found - Dish or vendor not found.
        '500':
          description: Internal Server Error - An unexpected error occurred.        
     
    put:
      tags:
        - Vendor
      summary: Update details of a dish
      operationId: updateDishDetails
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: dishId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dish'
            examples:
              example1:
                summary: Example dish update
                value:
                  name: "Updated Margherita Pizza"
                  imageLink: "http://example.com/new-images/margherita.jpg"
                  price: 13.00
                  allergens: ["Gluten", "Dairy"]
                  ingredients: ["Flour", "Tomatoes", "Mozzarella Cheese", "Basil", "Olive Oil"]
                  description: "Now the best in Hague too!"  
      responses:
        '200':
          description: Dish details updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
              examples:
                exampleResponse:
                  summary: Example response after updating a dish
                  value:
                    ID: "ccf1f2d7-8f7f-4044-9a30-16790a432977"
                    name: "Updated Margherita Pizza"
                    imageLink: "http://example.com/new-images/margherita.jpg"
                    price: 13.00
                    allergens: ["Gluten", "Dairy"]
                    ingredients: ["Flour", "Tomatoes", "Mozzarella Cheese", "Basil", "Olive Oil"]
                    description: "Now the best in Hague too!"
        '400':
          description: Bad Request - Incorrect dish details format or missing required fields.
        '401':
          description: Unauthorized - User is not a vendor/dish does not belong to vendor.
        '404':
          description: Not Found - Dish or vendor not found.
        '500':
          description: Internal Server Error - An unexpected error occurred.
  /vendor/{vendorId}/orders/{orderId}:
    get:
      tags:
        - Vendor
      summary: Get details of a specific order (including how much money is earned and special requirements)
      operationId: getOrderDetails
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Details of the specified order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Bad Request - Invalid request parameters.
        '401':
          description: Unauthorized - User is not a vendor/order does not belong to vendor.
        '404':
          description: Order or vendor not found.
        '500':
          description: Internal Server Error - An unexpected error occurred.
...
components:
  schemas:
    Order:
      type: object
      x-class-extra-annotation: '@javax.persistence.Entity @javax.persistence.Table(name = "customer_order")'
      properties:
        ID:
          type: string
          format: uuid
          x-field-extra-annotation: '@javax.persistence.Id @javax.persistence.GeneratedValue(generator = "UUID") @org.hibernate.annotations.GenericGenerator(name = "UUID", strategy = "org.hibernate.id.UUIDGenerator")'
        dishes:
          type: array
          items:
            $ref: '#/components/schemas/OrderedDish'
          x-field-extra-annotation: '@javax.persistence.OneToMany(cascade=javax.persistence.CascadeType.ALL)'
        location:
          $ref: '#/components/schemas/Address'
        specialRequirements:
          type: string
        status:
          $ref: '#/components/schemas/Status'
        totalPrice:
          type: number
          format: double
        orderTime:
          type: string
          format: date-time
          description: The timestamp when the order was created.  
        vendorId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
    Address:
      x-class-extra-annotation: '@javax.persistence.Embeddable'
      type: object
      properties:
        houseNumber:
          type: integer
        zip:
          type: string
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
...
    Dish:
      type: object
      x-class-extra-annotation: '@javax.persistence.Entity @javax.persistence.Table(name = "dish")'
      properties:
        ID:
          type: string
          format: uuid
          x-field-extra-annotation: '@javax.persistence.Id @javax.persistence.GeneratedValue(generator = "UUID") @org.hibernate.annotations.GenericGenerator(name = "UUID", strategy = "org.hibernate.id.UUIDGenerator")'
        name:
          type: string
        imageLink:
          type: string
        price:
          type: number
          format: double
        isDeleted:
          type: boolean
          default: false
        allergens:
          type: array
          items:
            type: string
          x-field-extra-annotation: '@javax.persistence.ElementCollection'
        ingredients:
          type: array
          items:
            type: string
          x-field-extra-annotation: '@javax.persistence.ElementCollection'
        description:
          type: string
        vendorId:
          type: string
          format: uuid
    OrderedDish:
      type: object
      x-class-extra-annotation: '@javax.persistence.Entity @javax.persistence.Table(name = "ordered_dish")'
      properties:
        id:
          type: string
          format: uuid
          x-field-extra-annotation: '@javax.persistence.Id @javax.persistence.GeneratedValue(generator = "UUID") @org.hibernate.annotations.GenericGenerator(name = "UUID", strategy = "org.hibernate.id.UUIDGenerator")'
        dish:
          allOf:
            - $ref: '#/components/schemas/Dish'
          x-field-extra-annotation: '@javax.persistence.ManyToOne'
        quantity:
          type: integer
          default: 0
    Status:
      type: string
      enum:
        - pending
        - accepted
        - rejected
        - preparing
        - givenToCourier
        - onTransit
        - delivered
...